Add files via upload
Write an ALP that will examine a string of size 20 bytes stored in location exm1. It counts the number of a times a certain character (small/capital) occurs and stores that count.

The count must be stored in memory location that has a label made up of the first 3 characters of your first name and last two numbers of your id no. (e.g. Anupama ID no 2019AAPS0014G - label will be anu14)

The character that you will search for depends on the first character of your first name for e.g. if your name starts with a 'A'(Anupama) you will search for 'Z', if your name starts with 'B' (Betty) you will search for 'Y', If your name starts with a  'C' (Carol) you will search for 'X', if your name starts with 'Z' (Zeena) you will search for 'A' , if your name starts with 'Y' (Yuvaraj) you will search for B and so on.
If your ID no is even for instance 2019A70018G - then you will search for small characters, else you will search for capital letters
You have store the starting letter of your first name in small alphabets at displacement 'name1' if you have an even id and in caps if you have an odd id.
The character to be searched for should be derived from the value store in 'name1'
You cannot store alphabets a to z in memory