Update TLabTest1.asm
There is a string made only of alphabets (small and caps) of size 30 in decimal stored at location sttr. Write an ALP that will examine each character and if it is in small alphabets, capitalize it, and replace it in the same location. You can assume that there are no special characters or numbers including spaces.