Add files via upload
Display the following pattern on the top left of the screen. The area of the outermost rectangle is 150 x 250

To do this - first display red rectangle of size 150 x250 at pixel 0,0. Then go to row 20 x 20 and display a green rectangle.

The size of green rectangle will be 110x 210.

keep repeating this - with alternate green and red rectangles till you get the required pattern.